requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
discord.py==2.6.4
distro==1.9.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
multidict==6.7.1
openai==2.15.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
PyJWT==2.10.1
PyMySQL==1.1.2
PyNaCl==1.6.2
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.46
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
